EES-4930 Move IApiSubscriptionTableStorageService/ApiSubscriptionTableStorageService from Notifier.Repositories to Notifier.Services
