+++ src/GovUk.Education.ExploreEducationStatistics.Notifier/Repositories/ApiSubscriptionRepository.cs
@@ -8,6 +8,7 @@ using Azure.Data.Tables;
 using GovUk.Education.ExploreEducationStatistics.Notifier.Configuration;
 using GovUk.Education.ExploreEducationStatistics.Notifier.Model;
 using GovUk.Education.ExploreEducationStatistics.Notifier.Repositories.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Notifier.Services.Interfaces;
 using Microsoft.Extensions.Options;
 
 namespace GovUk.Education.ExploreEducationStatistics.Notifier.Repositories;
