+++ src/GovUk.Education.ExploreEducationStatistics.Notifier/Services/ApiSubscriptionTableStorageService.cs
@@ -1,9 +1,9 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Services;
 using GovUk.Education.ExploreEducationStatistics.Notifier.Configuration;
-using GovUk.Education.ExploreEducationStatistics.Notifier.Repositories.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Notifier.Services.Interfaces;
 using Microsoft.Extensions.Options;
 
-namespace GovUk.Education.ExploreEducationStatistics.Notifier.Repositories;
+namespace GovUk.Education.ExploreEducationStatistics.Notifier.Services;
 
 internal class ApiSubscriptionTableStorageService(IOptions<AppSettingsOptions> appSettingsOptions)
     : DataTableStorageService(appSettingsOptions.Value.NotifierStorageConnectionString),
