+++ src/GovUk.Education.ExploreEducationStatistics.Notifier/Services/Interfaces/IApiSubscriptionTableStorageService.cs
@@ -1,5 +1,5 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Services.Interfaces;
 
-namespace GovUk.Education.ExploreEducationStatistics.Notifier.Repositories.Interfaces;
+namespace GovUk.Education.ExploreEducationStatistics.Notifier.Services.Interfaces;
 
 public interface IApiSubscriptionTableStorageService : IDataTableStorageService;
